+++ src/app.module.ts
@@ -17,6 +17,7 @@ import {
   OpenAIConfig,
   PizzaConfig,
   HacktoberfestConfig,
+  TierConfig,
 } from "./config";
 import { RepoModule } from "./repo/repo.module";
 import { HealthModule } from "./health/health.module";
@@ -77,6 +78,7 @@ import { CouponModule } from "./coupon/coupon.module";
 import { DbUserList } from "./user-lists/entities/user-list.entity";
 import { DbUserListContributor } from "./user-lists/entities/user-list-contributor.entity";
 import { DbCoupon } from "./coupon/entities/coupon.entity";
+import { LogModule } from "./log/log.module";
 
 @Module({
   imports: [
@@ -90,6 +92,7 @@ import { DbCoupon } from "./coupon/entities/coupon.entity";
         OpenAIConfig,
         PizzaConfig,
         HacktoberfestConfig,
+        TierConfig,
       ],
       isGlobal: true,
     }),
@@ -223,6 +226,7 @@ import { DbCoupon } from "./coupon/entities/coupon.entity";
     PizzaOvenModule,
     UserListModule,
     CouponModule,
+    LogModule,
   ],
   providers: [],
 })