+++ src/pull-requests/dtos/pull-request-contributor.dto.ts
@@ -7,6 +7,12 @@ export class DbPullRequestContributor {
   })
   public author_login: string;
 
+  @ApiProperty({
+    description: "Pull request author id",
+    example: 1,
+  })
+  public user_id: number;
+
   @ApiPropertyOptional({
     description: "Timestamp representing pr last update",
     example: "2022-08-28 22:04:29.000000",