+++ src/user-lists/dtos/filter-contributors.dto.ts
@@ -1,6 +1,6 @@
 import { ApiPropertyOptional } from "@nestjs/swagger";
-import { Type } from "class-transformer";
-import { IsInt, Min, IsOptional, Max, IsString } from "class-validator";
+import { Transform, Type } from "class-transformer";
+import { IsInt, Min, IsOptional, Max, IsString, IsArray } from "class-validator";
 
 export class FilterListContributorsDto {
   @ApiPropertyOptional({
@@ -28,12 +28,16 @@ export class FilterListContributorsDto {
   readonly limit?: number = 10;
 
   @ApiPropertyOptional({
+    isArray: true,
     type: "string",
-    example: "Denver, Colorado",
+    example: ["Denver, Colorado", "Germany"],
   })
-  @IsString()
+  @IsArray()
+  @IsString({ each: true })
+  @Type(() => String)
+  @Transform(({ value }) => (Array.isArray(value) ? value : Array(value)) as string[], { toClassOnly: true })
   @IsOptional()
-  location?: string;
+  location?: string[];
 
   @ApiPropertyOptional({
     type: "string",
@@ -44,12 +48,16 @@ export class FilterListContributorsDto {
   contributor?: string;
 
   @ApiPropertyOptional({
+    isArray: true,
     type: "string",
-    example: "Mountain Standard Time",
+    example: ["Mountain Standard Time", "UTC"],
   })
-  @IsString()
+  @IsArray()
+  @IsString({ each: true })
+  @Type(() => String)
+  @Transform(({ value }) => (Array.isArray(value) ? value : Array(value)) as string[], { toClassOnly: true })
   @IsOptional()
-  timezone?: string;
+  timezone?: string[];
 
   @ApiPropertyOptional({
     type: "integer",