+++ src/user/dtos/contributor-prs.dto.ts
@@ -1,0 +1,16 @@
+import { ApiPropertyOptional } from "@nestjs/swagger";
+import { IsEnum, IsOptional } from "class-validator";
+
+import { PageOptionsDto } from "../../common/dtos/page-options.dto";
+
+export enum RangeTypeEnum {
+  All = "all",
+  Recent = "recent",
+}
+
+export class ContributorPullRequestsDto extends PageOptionsDto {
+  @ApiPropertyOptional({ enum: RangeTypeEnum, enumName: "OrderDirectionEnum", default: RangeTypeEnum.Recent })
+  @IsEnum(RangeTypeEnum)
+  @IsOptional()
+  rangeType?: RangeTypeEnum = RangeTypeEnum.Recent;
+}