+++ src/user/entities/user-collaboration.entity.ts
@@ -42,7 +42,7 @@ export class DbUserCollaboration extends BaseEntity {
   public request_user_id?: number;
 
   @ApiHideProperty()
-  @ManyToOne(() => DbUser, (user) => user.collaborations)
+  @ManyToOne(() => DbUser, (user) => user.collaborations, { onDelete: "CASCADE" })
   @JoinColumn({
     name: "user_id",
     referencedColumnName: "id",
@@ -50,7 +50,7 @@ export class DbUserCollaboration extends BaseEntity {
   public user: Relation<DbUser>;
 
   @ApiHideProperty()
-  @ManyToOne(() => DbUser, (user) => user.request_collaborations)
+  @ManyToOne(() => DbUser, (user) => user.request_collaborations, { onDelete: "CASCADE" })
   @JoinColumn({
     name: "request_user_id",
     referencedColumnName: "id",